feat(circle): refactor circle detail pages
