+++ src/views/Circle/Broadcast/styles.css
@@ -1,3 +1,15 @@
+.container {
+  padding-top: var(--spacing-x-tight);
+  padding-bottom: var(--spacing-xxx-loose);
+  background: var(--color-white);
+  border-radius: 1.25rem 1.25rem 0 0;
+
+  @media (--sm-up) {
+    padding-top: 0;
+    padding-bottom: var(--spacing-xx-loose);
+  }
+}
+
 .broadcast {
   padding: 0 var(--spacing-base);
   margin-top: var(--spacing-x-tight);
