+++ src/views/Circle/Works/gql.ts
@@ -2,8 +2,6 @@ import gql from 'graphql-tag'
 
 import { ArticleDigestFeed } from '~/components'
 
-import SubscriptionBanner from '../SubscriptionBanner'
-
 export const CIRCLE_WORKS_PUBLIC = gql`
   query CircleWorksPublic($name: String!, $after: String) {
     circle(input: { name: $name }) {
@@ -22,21 +20,16 @@ export const CIRCLE_WORKS_PUBLIC = gql`
           }
         }
       }
-      ...SubscriptionBannerCirclePublic
-      ...SubscriptionBannerCirclePrivate
     }
   }
   ${ArticleDigestFeed.fragments.article.public}
   ${ArticleDigestFeed.fragments.article.private}
-  ${SubscriptionBanner.fragments.circle.public}
-  ${SubscriptionBanner.fragments.circle.private}
 `
 
 export const CIRCLE_WORKS_PRIVATE = gql`
   query CircleWorksPrivate($name: String!, $ids: [ID!]!) {
     circle(input: { name: $name }) {
       id
-      ...SubscriptionBannerCirclePrivate
     }
     nodes(input: { ids: $ids }) {
       id
@@ -46,5 +39,4 @@ export const CIRCLE_WORKS_PRIVATE = gql`
     }
   }
   ${ArticleDigestFeed.fragments.article.private}
-  ${SubscriptionBanner.fragments.circle.private}
 `
