+++ src/views/Circle/Works/index.tsx
@@ -1,180 +1,18 @@
-import { useContext, useEffect, useState } from 'react'
+import { Layout } from '~/components'
 
-import {
-  ArticleDigestFeed,
-  EmptyArticle,
-  InfiniteScroll,
-  List,
-  QueryError,
-  Spinner,
-  SubscribeCircleDialog,
-  Throw404,
-  useEventListener,
-  usePublicQuery,
-  usePullToRefresh,
-  useRoute,
-  ViewerContext,
-} from '~/components'
+import Profile from '../Profile'
+import styles from './styles.css'
+import Works from './Works'
 
-import { REFETCH_CIRCLE_DETAIL_ARTICLES } from '~/common/enums'
-import { analytics, mergeConnections } from '~/common/utils'
+const CircleWorks = () => (
+  <Layout.Main bgColor="grey-lighter">
+    <Profile />
 
-import CircleDetailContainer from '../Detail'
-import SubscriptionBanner from '../SubscriptionBanner'
-import { CIRCLE_WORKS_PRIVATE, CIRCLE_WORKS_PUBLIC } from './gql'
-
-import { CircleWorksPublic } from './__generated__/CircleWorksPublic'
-
-const Works = () => {
-  const { getQuery } = useRoute()
-  const viewer = useContext(ViewerContext)
-  const name = getQuery('name')
-
-  /**
-   * Data Fetching
-   */
-  // public data
-  const {
-    data,
-    loading,
-    error,
-    fetchMore,
-    refetch: refetchPublic,
-    client,
-  } = usePublicQuery<CircleWorksPublic>(CIRCLE_WORKS_PUBLIC, {
-    variables: { name },
-  })
-
-  // pagination
-  const connectionPath = 'circle.articles'
-  const circle = data?.circle
-  const circleId = circle?.id
-  const { edges, pageInfo } = circle?.articles || {}
-
-  // private data
-  const [privateFetched, setPrivateFetched] = useState(false)
-  const loadPrivate = async (publicData?: CircleWorksPublic) => {
-    if (!viewer.isAuthed || !publicData) {
-      return
-    }
-
-    const publiceEdges = publicData.circle?.articles.edges || []
-
-    const publicIds = publiceEdges.map(({ node }) => node.id)
-
-    await client.query({
-      query: CIRCLE_WORKS_PRIVATE,
-      fetchPolicy: 'network-only',
-      variables: { name, ids: publicIds },
-    })
-
-    setPrivateFetched(true)
-  }
-
-  // fetch private data
-  useEffect(() => {
-    if (!circleId) {
-      return
-    }
-
-    if (viewer.isAuthed) {
-      loadPrivate(data)
-    } else if (viewer.privateFetched) {
-      setPrivateFetched(true)
-    }
-  }, [circleId, viewer.privateFetched])
-
-  // load next page
-  const loadMore = async () => {
-    analytics.trackEvent('load_more', {
-      type: 'circle_detail',
-      location: edges?.length || 0,
-    })
-
-    const { data: newData } = await fetchMore({
-      variables: {
-        after: pageInfo?.endCursor,
-      },
-      updateQuery: (previousResult, { fetchMoreResult }) =>
-        mergeConnections({
-          oldData: previousResult,
-          newData: fetchMoreResult,
-          path: connectionPath,
-        }),
-    })
-
-    loadPrivate(newData)
-  }
-
-  // refetch, sync & pull to refresh
-  const refetch = async () => {
-    const { data: newData } = await refetchPublic()
-    loadPrivate(newData)
-  }
-
-  useEventListener(REFETCH_CIRCLE_DETAIL_ARTICLES, refetch)
-  usePullToRefresh.Handler(refetch)
-
-  /**
-   * Render
-   */
-  if (loading) {
-    return <Spinner />
-  }
-
-  if (error) {
-    return <QueryError error={error} />
-  }
-
-  if (!circle || !pageInfo) {
-    return <Throw404 />
-  }
-
-  return (
-    <section className="works">
-      {!edges || (edges.length <= 0 && <EmptyArticle />)}
-
-      <InfiniteScroll hasNextPage={pageInfo.hasNextPage} loadMore={loadMore}>
-        <List>
-          {(edges || []).map(({ node, cursor }, i) => (
-            <List.Item key={cursor}>
-              <ArticleDigestFeed
-                article={node}
-                hasCircle={false}
-                onClick={() =>
-                  analytics.trackEvent('click_feed', {
-                    type: 'circle_detail',
-                    contentType: 'article',
-                    styleType: 'title',
-                    location: i,
-                  })
-                }
-                onClickAuthor={() => {
-                  analytics.trackEvent('click_feed', {
-                    type: 'circle_detail',
-                    contentType: 'user',
-                    styleType: 'subtitle',
-                    location: i,
-                  })
-                }}
-              />
-            </List.Item>
-          ))}
-        </List>
-      </InfiniteScroll>
-
-      <SubscribeCircleDialog circle={circle} />
-      {privateFetched && <SubscriptionBanner circle={circle} />}
-    </section>
-  )
-}
-
-const CircleDetailWorks = () => {
-  return (
-    <CircleDetailContainer>
+    <section className="container">
       <Works />
-    </CircleDetailContainer>
-  )
-}
+      <style jsx>{styles}</style>
+    </section>
+  </Layout.Main>
+)
 
-export default CircleDetailWorks
+export default CircleWorks
