+++ src/views/Circle/Works/styles.css
@@ -1,4 +1,4 @@
-.content {
+.container {
   padding-top: var(--spacing-x-tight);
   padding-bottom: var(--spacing-xxx-loose);
   background: var(--color-white);
